refactor(#3038): move tmux watcher prompt observation
